clean ups, tests
