+++ backend/src/Designer/Controllers/RepositoryController.cs
@@ -185,12 +185,16 @@ namespace Altinn.Studio.Designer.Controllers
                 return BadRequest($"{repository} is an invalid repository name.");
             }
 
-            var config = new ServiceConfiguration { RepositoryName = repository, ServiceName = new LocalizedString()
+            var config = new ServiceConfiguration
             {
-                Nb = repository,
-                Nn = "",
-                En = ""
-            } };
+                RepositoryName = repository,
+                ServiceName = new LocalizedString()
+                {
+                    Nb = repository,
+                    Nn = "",
+                    En = ""
+                }
+            };
 
             var repositoryResult = await _repository.CreateService(org, config);
             if (repositoryResult.RepositoryCreatedStatus == HttpStatusCode.Created)
