+++ backend/src/Designer/Converters/LocalizedStringConverter.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using Altinn.Studio.Designer.Models;
 
 public class LocalizedStringConverter : JsonConverter<LocalizedString>
 {
