+++ backend/src/Designer/Enums/AvailableForType.cs
@@ -1,0 +1,10 @@
+namespace Altinn.Studio.Designer.Enums;
+
+public enum AvailableForType
+{
+    PrivatePerson,
+    LegalEntityEnterprise,
+    Company,
+    BankruptcyEstate,
+    SelfRegisteredUser
+}
