+++ backend/src/Designer/Enums/ServiceStatus.cs
@@ -1,0 +1,9 @@
+namespace Altinn.Studio.Designer.Enums;
+
+public enum ServiceStatus
+{
+    Completed,
+    Deprecated,
+    UnderDevelopment,
+    Withdrawn
+}
