+++ backend/src/Designer/Models/LocalizedString.cs
@@ -1,0 +1,8 @@
+namespace Altinn.Studio.Designer.Models;
+
+public class LocalizedString
+{
+    public string Nb { get; set; }
+    public string Nn { get; set; }
+    public string En { get; set; }
+}
