+++ backend/tests/Designer.Tests/Services/RepositorySITests.cs
@@ -119,12 +119,16 @@ namespace Designer.Tests.Services
 
             try
             {
-                await repositoryService.CreateService(org, new ServiceConfiguration() { RepositoryName = repositoryName, ServiceName = new LocalizedString()
+                await repositoryService.CreateService(org, new ServiceConfiguration()
                 {
-                    Nb = repositoryName,
-                    Nn = repositoryName,
-                    En = repositoryName
-                } });
+                    RepositoryName = repositoryName,
+                    ServiceName = new LocalizedString()
+                    {
+                        Nb = repositoryName,
+                        Nn = repositoryName,
+                        En = repositoryName
+                    }
+                });
                 var altinnStudioSettings = await new AltinnGitRepositoryFactory(repositoriesRootDirectory).GetAltinnGitRepository(org, repositoryName, developer).GetAltinnStudioSettings();
                 Assert.Equal(AltinnRepositoryType.App, altinnStudioSettings.RepoType);
             }
